@thunderstore/sertra: rename file for consistency
Refs TS-670
