+++ apps/sertra/components/ListingAttributes.tsx
